+++ packages/viewer/web/src/getPluginViewers.jsx
@@ -68,14 +68,12 @@ class PluginViewer extends PureComponent {
 
   innerRCV = ({ contentState, textAlignment, direction }) => {
     const { innerRCEViewerProps } = this.props;
-    const config = this.removeExpand(innerRCEViewerProps.config);
     return (
       <RichContentViewer
         initialState={contentState}
         textAlignment={textAlignment}
         direction={direction}
         {...innerRCEViewerProps}
-        config={config}
       />
     );
   };
@@ -93,6 +91,7 @@ class PluginViewer extends PureComponent {
       context,
       blockIndex,
       SpoilerViewerWrapper,
+      blockKey,
     } = this.props;
     const { component: Component, elementType } = pluginComponent;
     const { container } = pluginComponent.classNameStrategies || {};
@@ -106,6 +105,7 @@ class PluginViewer extends PureComponent {
       entityIndex,
       ...context,
       innerRCV: this.innerRCV,
+      blockKey,
     };
 
     if (Component) {
@@ -227,6 +227,7 @@ PluginViewer.propTypes = {
   }).isRequired,
   innerRCEViewerProps: PropTypes.object,
   blockIndex: PropTypes.number,
+  blockKey: PropTypes.string,
 };
 
 PluginViewer.defaultProps = {
@@ -263,6 +264,7 @@ const getPluginViewers = (
           innerRCEViewerProps={innerRCEViewerProps}
           SpoilerViewerWrapper={SpoilerViewerWrapper}
           withHorizontalScroll
+          blockKey={block.key}
         >
           {isInline ? children : null}
         </PluginViewer>